Improved chart component (#26)
* created new functions in API


* Moved router push to context


* changes in cookie provider, setcookie created


* tested themes


* topbar color changed


* context api modified


* added chart to results


* fixed steps sequence and theme button


* modified chart,, component created


---------
